(Temporarily) enable logging of server and curl for troubleshooting
I first had issues with curl, because inside GVIM, the $no_proxy variable wasn't set somehow, and curl attempted to contact the localhost URL via the proxy, which obviously doesn't work.
Extract a new g:instant_markdown_logfile configuration that defaults to the (platform-specific) null device, and append (>> instead of overwriting via >). If troubleshooting is needed, the config can be set to a log filespec.